+++ .playground/i18n.config.ts
@@ -1,0 +1,3 @@
+import { I18N_VUE_CONFIG } from '@dargmuesli/nuxt-vio/utils/constants'
+
+export default defineI18nConfig(() => I18N_VUE_CONFIG)