+++ .playground/pages/index.vue
@@ -1,6 +1,17 @@
 <template>
   <div class="flex justify-center items-center vio-prose">
-    <h1>Test</h1>
-    <VioLink to="#">Test</VioLink>
+    <h1>{{ t('title') }}</h1>
+    <VioLink to="/">{{ t('globalLoading') }}</VioLink>
   </div>
 </template>
+
+<script setup lang="ts">
+const { t } = useI18n()
+</script>
+
+<i18n lang="yml">
+de:
+  title: Titel
+en:
+  title: Title
+</i18n>