+++ app.config.ts
@@ -1,7 +1,5 @@
 import { useSeoMeta } from '@unhead/vue'
 
-import { SITE_NAME } from './utils/constants'
-
 export default defineAppConfig({
   legalNotice: undefined as
     | {
@@ -47,7 +45,9 @@ export default defineAppConfig({
         }
       }
     | undefined,
-  seoMeta: undefined as Parameters<typeof useSeoMeta>[0] | undefined,
+  seoMeta: {
+    twitterSite: '@dargmuesli',
+  } as Parameters<typeof useSeoMeta>[0] | undefined,
   siteName: SITE_NAME,
-  themeColor: undefined as string | undefined,
+  themeColor: '#202020' as string | undefined,
 })