+++ tailwind.config.ts
@@ -7,7 +7,7 @@ const heading = (theme: PluginAPI['theme']) =>
     fontWeight: theme('fontWeight.bold'),
     overflow: 'hidden',
     textOverflow: 'ellipsis',
-  } as Record<string, string>)
+  }) as Record<string, string>
 
 const gray = colors.gray // or slate, zinc, neutral, stone